+++ NEWS.md
@@ -1,5 +1,15 @@
 # lazytrade (development version)
 
+# lazytrade 0.3.6
+
+# Version 0.3.6
+
+## Changes
+
+* removal of duplicate rows in the `aml_collect_data()` function
+* added a test for `to_m()` function
+* add warning to indicate if 'zeroes' are present in the data file column(s)
+* removing dependency from tidyverse
 
 # lazytrade 0.3.5